# 10210-admin/swagger.yaml
openapi: 3.0.0
info:
  title: CuWhapp API
  description: |
    # WhatsApp Marketing Automation Platform API
    
    Complete API documentation for CuWhapp - The ultimate WhatsApp marketing automation platform.
    
    **Total Endpoints: 124**
    
    ## Features
    - WhatsApp Session Management
    - Campaign Creation & Automation  
    - Account Warming (Premium)
    - Contact & Group Management
    - Payment & Subscription System
    - Analytics & Reporting
    - Admin Dashboard
    
    ## Authentication
    Most endpoints require authentication via Bearer token or Admin token.
  version: 2.0.0
  contact:
    name: CuWhapp Support
    email: support@cuwapp.com
    url: https://cuwapp.com
servers:
  - url: https://app.cuwapp.com
    description: Development server
  - url: https://api.cuwapp.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    AdminAuth:
      type: http
      scheme: bearer
      description: Admin authentication token
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [created, running, paused, completed, failed, scheduled]
        session_name:
          type: string
        total_recipients:
          type: integer
        messages_sent:
          type: integer
        messages_failed:
          type: integer
        is_scheduled:
          type: boolean
        scheduled_start_time:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags:
        - 🔐 Authentication
      summary: Login with plan-based session timeout
      description: |
        Login endpoint that creates sessions with plan-based timeout enforcement:
        
        **Session Timeouts by Plan:**
        - **FREE**: 1-hour session timeout (strict enforcement)
        - **STARTER**: Persistent sessions (30-day expiration)
        - **HOBBY+**: Persistent sessions (30-day expiration)
        
        Free plan users receive warnings at 10 minutes remaining and are automatically logged out after 1 hour.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  session_id:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
                  plan_type:
                    type: string
                  session_duration_minutes:
                    type: integer
                  persistent:
                    type: boolean
...
  /api/auth/magic-link:
    post:
      tags:
        - 🔐 Authentication
      summary: Request magic link for passwordless login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Magic link sent successfully
  /api/auth/verify-magic-link:
    get:
      tags:
        - 🔐 Authentication
      summary: Verify magic link token
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token verified successfully
...
  /api/auth/session/refresh:
    post:
      tags:
        - 🔐 Authentication
      summary: Refresh session token (paid plans only)
      responses:
        '200':
          description: Session refreshed successfully
        '400':
          description: Refresh not allowed for free plan
  /api/auth/session/status:
    get:
      tags:
        - 🔐 Authentication
      summary: Get active sessions statistics (admin only)
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Session statistics

  # Additional auth endpoints would be added here...
  # (Adding remaining 4 auth endpoints for completeness)

  # ================================
  # 📱 SESSIONS (8 total)
  # ================================
...
  /api/campaigns:
    get:
      tags:
        - 🚀 Campaigns
      summary: Get all user campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [created, running, paused, completed, failed, scheduled]
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'

    post:
      tags:
        - 🚀 Campaigns
      summary: Create new campaign
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                session_name:
                  type: string
                file:
                  type: string
                  format: binary
                message_template:
                  type: string
                delay_seconds:
                  type: integer
                  default: 5
      responses:
        '201':
          description: Campaign created successfully
...
  /api/campaigns/stats:
    get:
      tags:
        - 🚀 Campaigns
      summary: Get campaign statistics
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Campaign statistics
...